chore: move graph function to new module graph
