+++ dashboard/pages/04_🎓_tesis.py
@@ -6,7 +6,8 @@ from typing import List
 import pandas as pd
 import streamlit as st
 from models import Thesis
-from modules.utils import generate_widget_key, build_advisors_graph
+from modules.utils import generate_widget_key
+from modules.graph import build_advisors_graph
 
 st.set_page_config(page_title="MatCom Dashboard - Tesis", page_icon="🎓", layout="wide")
